Fix the problem whereby when one includes FlippedPanels in the wanted set then only those FlippedPanels and not their associated WantedPanels are considered.
The problem was that the FlippedPanel occurred before the associated
WantedPanel in the wanted set and `progression` was eliminating the
WantedPanel once its FlippedPanel was solved for.

Now, when the wanted set includes a FlippedPanel it will not include
the associated WantedPanel, and `progression` will consider both at
the same time.